requirements.txt
python-dotenv>=1.0.0
langgraph>=0.2.0
langchain-anthropic>=0.1.0
anthropic>=0.40.0
pydantic>=2.0